Change Datatype and validations

+++ ap_express_server/src/main/java/com/ap_express_server/repository/accountType/AccountTypeRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 import java.util.List;
 
-public interface AccountTypeRepository extends JpaRepository<AccountType, Long> {
+public interface AccountTypeRepository extends JpaRepository<AccountType, Integer> {
 
     /**
      * thisa method query from account_type table and get account type list
