+++ ap_express_server/src/main/java/com/ap_express_server/repository/payment/PaymentMethodRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 import java.util.List;
 
-public interface PaymentMethodRepository extends JpaRepository<PaymentMethod, Long> {
+public interface PaymentMethodRepository extends JpaRepository<PaymentMethod, Integer> {
     @Query("SELECT new com.ap_express_server.models.dropdown.DropDownDto(pMethod.id, pMethod.name) FROM PaymentMethod pMethod")
     List<DropDownDto> getPaymentMethodList();
 }
